Se crea clase extractFile
